requirement.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
